Add a new FilterAndSearch strategy
This strategy allows you to both filter and search while within a
certain data size limit.

The basic algorithm is:
  - We apply our filters and get ALL of the ids from that scope (with no
      pagination applied)
  - We apply our search and get ALL of the ids from that search (with no
      pagination applied)
  - We take the intersection of those two arrays of ids
  - We do a final query against the original scope where the id is in
  the above intersection

The reason we have to get the intersection of the entirety of both sets
is that otherwise we do not have an accurate COUNT that is used for
paginating. This could potentially be SLOW as you could be doing a large
query and search, depending on your problem size. There is another
possible strategy we will implement for problem sizes over a certain
threshold that should be more performant for those large sizes, but will
NOT produce an accurate COUNT. So any pagination you're doing will need
to take into account that you won't necessarily know what the final page
is. That's not a problem with this commit, however.