Retrieve all search IDs and THEN select them from the scope, and then filter
This resolves a problem where having greater than
`default_max_filter_and_search_page` possibilities could return
inaccurate results. This is because our scope and our search could
potentially be searching a different 10,000 objects, depending on the
search and the filter options. This changes ensures that the up to
10,000 IDs that we are filtering are the exact 10,000 IDs match our
search.