Return an empty array when has_many association is nil
There is a bug that if a dynamic/lookup assocation in a presenter
returns nil, it is always treated as a has_one relationship. Thus the
foreign key is incorrect, e.g., `workspace_id` instead of
`workspace_ids`. Further the foreign key value to be inconsistent with
empty `Array`s and `ActiveRecord::Relation`s. So we now ensure we return
an empty array if it is nil